+++ obsi.lua
@@ -59,7 +59,7 @@ obsi.mouse, mouseDown, mouseUp, mouseMove = require("mouse")()
 obsi.audio, soundLoop = require("audio")(gamePath)
 obsi.state = require("state")
 obsi.debug = false
-obsi.version = "1.4.1"
+obsi.version = "1.4.2"
 -- obsi.debugger = peripheral.find("debugger") or (periphemu and periphemu.create("right", "debugger") and peripheral.find("debugger"))
 
 local chunk, err = loadfile(fs.combine(gamePath, "main.lua"), "bt", env)